+++ 1.20/Fabric/src/main/java/fuzs/thinair/core/FabricAbstractions.java
@@ -8,16 +8,14 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 
-import java.util.Optional;
-
 public class FabricAbstractions implements CommonAbstractions {
 
     @Override
-    public Optional<ItemStack> findEquippedItem(LivingEntity entity, TagKey<Item> tagKey) {
+    public ItemStack findEquippedItem(LivingEntity entity, TagKey<Item> tagKey) {
         if (ModLoaderEnvironment.INSTANCE.isModLoaded("trinkets")) {
             return TrinketsApi.getTrinketComponent(entity).flatMap(component -> component.getEquipped(itemStack -> itemStack.is(tagKey))
                     .stream().findFirst().map(Tuple::getB))
-                    .or(() -> CommonAbstractions.super.findEquippedItem(entity, tagKey));
+                    .orElseGet(() -> CommonAbstractions.super.findEquippedItem(entity, tagKey));
         } else {
             return CommonAbstractions.super.findEquippedItem(entity, tagKey);
         }